+++ include/openni2_camera/openni2_frame_listener.h
@@ -38,7 +38,7 @@
 
 #include <vector>
 
-#include "OpenNI.h"
+#include "OpenNI-2/OpenNI.h"
 
 namespace openni2_wrapper
 {