+++ src/openni2_device.cpp
@@ -29,7 +29,8 @@
  *      Author: Julius Kammerl (jkammerl@willowgarage.com)
  */
 
-#include "OpenNI.h"
+#include "OpenNI-2/OpenNI.h"
+#include "NiTE-2/NiTE.h"
 
 #include <boost/lexical_cast.hpp>
 #include <boost/algorithm/string/replace.hpp>
@@ -38,6 +39,7 @@
 #include "openni2_camera/openni2_exception.h"
 #include "openni2_camera/openni2_convert.h"
 #include "openni2_camera/openni2_frame_listener.h"
+#include "openni2_camera/nite2_hand_tracker_frame_listener.h"
 
 #include <boost/shared_ptr.hpp>
 #include <boost/make_shared.hpp>
@@ -52,12 +54,23 @@ OpenNI2Device::OpenNI2Device(const std::string& device_URI) throw (OpenNI2Except
     ir_video_started_(false),
     color_video_started_(false),
     depth_video_started_(false),
+    hand_tracker_started_(false),
+    has_nite_(false),
     image_registration_activated_(false),
     use_device_time_(false)
-{
-  openni::Status rc = openni::OpenNI::initialize();
+{  
+  openni::Status rc = openni::STATUS_OK;
+
+  rc = openni::OpenNI::initialize();
   if (rc != openni::STATUS_OK)
     THROW_OPENNI_EXCEPTION("Initialize failed\n%s\n", openni::OpenNI::getExtendedError());
+
+  //Initialize NiTE2
+  nite::Status niteRc = nite::NiTE::initialize();
+  if (niteRc != nite::STATUS_OK)
+    std::cout << "NiTE2 initialization failed. Hand tracking won't be available" << std::endl;
+  else
+    has_nite_ = true;
 
   openni_device_ = boost::make_shared<openni::Device>();
 
@@ -79,7 +92,8 @@ OpenNI2Device::OpenNI2Device(const std::string& device_URI) throw (OpenNI2Except
   ir_frame_listener = boost::make_shared<OpenNI2FrameListener>();
   color_frame_listener = boost::make_shared<OpenNI2FrameListener>();
   depth_frame_listener = boost::make_shared<OpenNI2FrameListener>();
-
+  if ( has_nite_ )
+    hand_tracker_frame_listener_ = boost::make_shared<nite2_wrapper::NiTE2HandTrackerFrameListener>();
 }
 
 OpenNI2Device::~OpenNI2Device()
@@ -87,6 +101,9 @@ OpenNI2Device::~OpenNI2Device()
   stopAllStreams();
 
   shutdown();
+
+  if ( has_nite_ )
+    nite::NiTE::shutdown();
 
   openni_device_->close();
 }
@@ -238,7 +255,12 @@ bool OpenNI2Device::hasColorSensor() const
 
 bool OpenNI2Device::hasDepthSensor() const
 {
-  return openni_device_->hasSensor(openni::SENSOR_DEPTH);
+  return openni_device_->hasSensor(openni::SENSOR_DEPTH);    
+}
+
+bool OpenNI2Device::hasHandTracker() const
+{
+  return has_nite_;
 }
 
 void OpenNI2Device::startIRStream()
@@ -280,11 +302,29 @@ void OpenNI2Device::startDepthStream()
   }
 }
 
+void OpenNI2Device::startHandTracker()
+{
+  if ( !has_nite_ )
+      THROW_OPENNI_EXCEPTION("Not starting hand tracking because NiTE2 could not be initialized");
+
+  boost::shared_ptr<nite::HandTracker> hand_tracker = getHandTracker();
+
+  if ( hand_tracker )
+  {
+    hand_tracker->addNewFrameListener(hand_tracker_frame_listener_.get());
+    hand_tracker->startGestureDetection(nite::GESTURE_WAVE);
+    hand_tracker->startGestureDetection(nite::GESTURE_CLICK);
+    hand_tracker_started_ = true;
+  }
+
+}
+
 void OpenNI2Device::stopAllStreams()
 {
   stopIRStream();
   stopColorStream();
   stopDepthStream();
+  stopHandTracker();
 }
 
 void OpenNI2Device::stopIRStream()
@@ -321,6 +361,15 @@ void OpenNI2Device::stopDepthStream()
   }
 }
 
+void OpenNI2Device::stopHandTracker()
+{
+  if (hand_tracker_.get() != NULL )
+  {
+    hand_tracker_started_ = false;
+    hand_tracker_->destroy();
+  }
+}
+
 void OpenNI2Device::shutdown()
 {
   if (ir_video_stream_.get() != 0)
@@ -332,6 +381,8 @@ void OpenNI2Device::shutdown()
   if (depth_video_stream_.get() != 0)
     depth_video_stream_->destroy();
 
+  if ( hand_tracker_.get() != 0 )
+    hand_tracker_->destroy();
 }
 
 bool OpenNI2Device::isIRStreamStarted()
@@ -345,6 +396,10 @@ bool OpenNI2Device::isColorStreamStarted()
 bool OpenNI2Device::isDepthStreamStarted()
 {
   return depth_video_started_;
+}
+bool OpenNI2Device::isHandTrackerStarted()
+{
+  return hand_tracker_started_;
 }
 
 const std::vector<OpenNI2VideoMode>& OpenNI2Device::getSupportedIRVideoModes() const
@@ -611,6 +666,11 @@ void OpenNI2Device::setDepthFrameCallback(FrameCallbackFunction callback)
   depth_frame_listener->setCallback(callback);
 }
 
+void OpenNI2Device::setHandTrackerFrameCallback(nite2_wrapper::HandTrackerFrameCallbackFunction callback)
+{
+  hand_tracker_frame_listener_->setCallback(callback);
+}
+
 boost::shared_ptr<openni::VideoStream> OpenNI2Device::getIRVideoStream() const throw (OpenNI2Exception)
 {
   if (ir_video_stream_.get() == 0)
@@ -657,6 +717,27 @@ boost::shared_ptr<openni::VideoStream> OpenNI2Device::getDepthVideoStream() cons
     }
   }
   return depth_video_stream_;
+}
+
+boost::shared_ptr<nite::HandTracker> OpenNI2Device::getHandTracker() const throw (OpenNI2Exception)
+{
+  if ( !has_nite_ )
+    THROW_OPENNI_EXCEPTION("Not creating hand tracking because NiTE2 could not be initialized");
+
+  if ( openni_device_.get() == NULL )
+  {
+    THROW_OPENNI_EXCEPTION("Not creating hand tracking because device has not been initialized");
+  }
+
+  hand_tracker_.reset(new nite::HandTracker);
+
+  nite::Status niteRc = hand_tracker_->create(openni_device_.get());
+  if (niteRc != nite::STATUS_OK)
+  {
+    THROW_OPENNI_EXCEPTION("Couldn't create hand tracker");
+  }
+
+  return hand_tracker_;
 }
 
 std::ostream& operator <<(std::ostream& stream, const OpenNI2Device& device)
@@ -712,6 +793,8 @@ std::ostream& operator <<(std::ostream& stream, const OpenNI2Device& device)
   {
     stream << "No Depth sensor available" << std::endl;
   }
+
+  return stream;
 
 }