+++ src/openni2_frame_listener.cpp
@@ -28,7 +28,7 @@
  *
  *      Author: Julius Kammerl (jkammerl@willowgarage.com)
  */
-#include "OpenNI.h"
+#include "OpenNI-2/OpenNI.h"
 
 #include "openni2_camera/openni2_frame_listener.h"
 #include "openni2_camera/openni2_timer_filter.h"
@@ -57,6 +57,8 @@ bool OpenNI2FrameListener::setUseDeviceTimer(bool enable)
 
   if (user_device_timer_)
     timer_filter_->clear();
+
+  return true;
 }
 
 void OpenNI2FrameListener::onNewFrame(openni::VideoStream& stream)