+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog for package openni2_camera
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.1.3 (2014-05-14)
+------------------
+* Include local copy of headers and binaries of NiTE 2.2.0.10 and OpenNI 2.2.0.33
+* Integreate NiTE2 hand tracker in openni2_camera node
+* Add new topic ~<name>/gestures publishing pal_vision_msgs/Gesture when Waving or Click is detected
+* Contributors: Jordi Pages
+
 0.1.2 (2014-02-03)
 ------------------
 * Fix CMake error.